Refactor nav functionality, add tab switching
It's a lot easier to deal with the nav if it doesn't get replaced every
time a new tab is selected.

I was also working with React in another project, so I was initially
thinking about how to set it up that way, but without hooks, but that
would be too much work.